doc/readthedocs-pip-requirements.txt
numpydoc==0.4
numpy
ipython
matplotlib
scipy

# an unreleased commit that will eventually make the 3.1.0 release, with
# type annotations which work on cached_property
-e git+https://git@github.com/sphinx-doc/sphinx.git@5460ad6925199b57b27b7d059825d3560872edbb#egg=sphinx

ipykernel
nbsphinx
numba
sphinx_rtd_theme
recommonmark
sphinx-markdown-tables
notedown
semantic-version==2.6.0
releases>=1.6.1
sphinxcontrib-bibtex
nbsphinx-link

# the first release to support f-strings
pygments >= 2.5.0
